requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.29.2=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
attrs=23.2.0=pypi_0
av=12.0.0=pypi_0
backports-tarfile=1.2.0=pypi_0
bitsandbytes=0.43.0=pypi_0
blas=1.0=mkl
braceexpand=0.1.7=pypi_0
brotli-python=1.0.9=py310h6a678d5_7
bypy=1.8.5=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py310h06a4308_0
cffi=1.16.0=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
contourpy=1.2.1=pypi_0
cryptography=42.0.8=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
decorator=4.4.2=pypi_0
decord=0.6.0=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docutils=0.21.2=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.1=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=py310h06a4308_0
fonttools=4.51.0=pypi_0
fsspec=2023.10.0=py310h06a4308_0
ftfy=6.2.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
gradio=4.26.0=pypi_0
gradio-client=0.15.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.4=pypi_0
idna=3.4=py310h06a4308_0
imageio=2.34.0=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
img2txt-py=2.4=pypi_0
importlib-metadata=7.0.1=py310h06a4308_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
jaraco-classes=3.4.0=pypi_0
jaraco-context=5.3.0=pypi_0
jaraco-functools=4.0.1=pypi_0
jeepney=0.8.0=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.4.0=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
keyring=25.2.1=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
more-itertools=10.2.0=pypi_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.3=pypi_0
nh3=0.2.17=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.17.0=pypi_0
opencv-python=4.10.0.82=pypi_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.0=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=23.2=py310h06a4308_0
pandas=2.2.2=pypi_0
panopticapi=0.1=pypi_0
parameterized=0.9.0=pypi_0
peft=0.2.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py310h06a4308_0
pkginfo=1.11.0=pypi_0
portalocker=2.8.2=pypi_0
proglog=0.1.10=pypi_0
progressbar2=4.4.2=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
pycocoevalcap=1.2=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310h06a4308_0
pysrt=1.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
python-utils=3.8.2=pypi_0
pytorchvideo=0.1.5=pypi_0
pytube=15.0.0=pypi_0
pytubefix=5.1.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
readme-renderer=43.0=pypi_0
referencing=0.34.0=pypi_0
regex=2023.10.3=py310h5eee18b_0
requests=2.31.0=py310h06a4308_1
requests-toolbelt=1.0.0=pypi_0
rfc3986=2.0.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
ruff=0.3.5=pypi_0
safetensors=0.4.2=py310ha89cbab_0
scikit-image=0.23.1=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.0=pypi_0
secretstorage=3.3.3=pypi_0
selenium=4.21.0=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=1.45.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py310h06a4308_0
shapely=2.0.4=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soundfile=0.12.1=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.37.2=pypi_0
submitit=1.5.1=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.3.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.2=pypi_0
torchaudio=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
towhee=1.1.3=pypi_0
towhee-models=1.1.3=pypi_0
tqdm=4.65.0=py310h2f386ee_0
transformers=4.30.2=pypi_0
trio=0.25.1=pypi_0
trio-websocket=0.11.1=pypi_0
triton=2.2.0=pypi_0
twine=5.1.0=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.9.0=py310h06a4308_1
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024.1=pypi_0
urllib3=2.1.0=py310h06a4308_1
uvicorn=0.29.0=pypi_0
visual-genome=1.1.1=pypi_0
wandb=0.16.6=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.86=pypi_0
websockets=11.0.3=pypi_0
webvtt-py=0.4.6=pypi_0
wheel=0.41.2=py310h06a4308_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h5eee18b_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
